refactor: 설정 모달을 반응형 대응 가능한 공통 컴포넌트로 분리
