+++ frontend/src/router/router.tsx
@@ -4,7 +4,6 @@ import MainPage from '../pages/main/MainPage'
 import LibraryPage from '../pages/library/LibraryPage'
 import ReportPage from '../pages/report/ReportPage'
 import MyPage from '../pages/my/MyPage'
-import SettingPage from '../pages/setting/SettingPage'
 
 export const router = createBrowserRouter([
     {
@@ -26,10 +25,6 @@ export const router = createBrowserRouter([
             {
                 path: '/library',
                 element: <LibraryPage />,
-            },
-            {
-                path: '/setting',
-                element: <SettingPage />,
             },
         ],
     },
